Create react redux file structure and move App into components
